requirements.txt
Django==1.5.1
django-annoying